feat: add server side validation to delete routes

+++ src/routes/societies/[societyId]/+server.ts
@@ -10,6 +10,16 @@ export const DELETE: RequestHandler = async ({ locals, params }) => {
 		throw error(400, "Invalid parameters");
 	}
 
+	const society = await SocietyService.get(params.societyId);
+
+	if (!society) {
+		throw error(404, "Society not found");
+	}
+
+	if (society.creator !== locals.user.web3Address) {
+		throw error(403, "You are not allowed to delete this society");
+	}
+
 	const success = await SocietyService.delete(params.societyId);
 
 	return new Response(JSON.stringify({ success }), { status: 200 });
