+++ src/routes/societies/[societyId]/courses/[courseId]/modules/[moduleId]/+server.ts
@@ -1,5 +1,6 @@
 import { error, type RequestHandler } from "@sveltejs/kit";
 import ModuleService from "$lib/services/ModuleService";
+import CourseService from "$lib/services/CourseService";
 
 export const DELETE: RequestHandler = async ({ locals, params }) => {
 	if (!locals.user) {
@@ -8,6 +9,16 @@ export const DELETE: RequestHandler = async ({ locals, params }) => {
 
 	if (!params.societyId || !params.courseId || !params.moduleId) {
 		throw error(400, "Invalid parameters");
+	}
+
+	const course = await CourseService.get(params.societyId, params.courseId);
+
+	if (!course) {
+		throw error(404, "Course not found");
+	}
+
+	if (course.creator !== locals.user.web3Address) {
+		throw error(403, "You are not the owner of this course");
 	}
 
 	const success = await ModuleService.delete(
